Revert "Home and library pages"
