+++ package.json
@@ -54,6 +54,6 @@
     "basiclightbox": "^5.0.4",
     "core-js": "^3.6.5",
     "handlebars": "^4.7.6",
-    "tui-pagination": "^3.4.0"
+    "paginationjs": "^2.1.5"
   }
 }
