--- src/js/open-close-modal.js
@@ -1,9 +1,0 @@
-import refs from './refs';
-
-/* function openModal() {
-  refs.modalBackdropRef.classList.toggle('show');
-}
-
-function closeModal() {
-  refs.modalBackdropRef.classList.toggle('show');
-} */
