--- src/scss/components/_headerСhanging.scss
@@ -1,67 +1,0 @@
-.filmSearch {
-  display: block;
-  text-align: center;
-
-  &__input {
-    width: 280px;
-
-    background: transparent;
-    border: transparent;
-    border-bottom: 1px solid $white-color;
-    font-size: 14px;
-    line-height: 1.17;
-    color: $white-color;
-    transition: 250ms color cubic-bezier(0.4, 0, 0.2, 1);
-    &::placeholder {
-      color: $white-color;
-    }
-
-    &:focus {
-      outline: none;
-      border-bottom: 1px solid $orange-color;
-    }
-  }
-  &__input-error {
-    font-size: 14px;
-    line-height: 1.14;
-    color: $red-color;
-    // position: absolute;
-    // top: 225%;
-  }
-}
-.white {
-  color: $white-color !important;
-}
-.mylibrary {
-  display: block;
-  text-align: center;
-
-  &__button {
-    width: 132px;
-    padding-top: 14px;
-    padding-bottom: 12px;
-    &:not(:last-child) {
-      margin-right: 15px;
-    }
-    border: 1px solid $white-color;
-    box-sizing: border-box;
-    border-radius: 5px;
-    background-color: transparent;
-
-    font-weight: 500;
-    font-size: 12px;
-    line-height: 1.33;
-    color: $white-color;
-    &:hover {
-      background-color: $orange-second-color;
-    }
-  }
-}
-.current-mylibrary__button {
-  background-color: $orange-color;
-  border: 1px solid transparent;
-}
-
-.is__hidden {
-  display: none;
-}
