--- src/scss/components/_modal.scss
@@ -1,244 +1,0 @@
-h2,
-h3 {
-  margin: 0;
-}
-p {
-  margin: 0;
-}
-
-ul {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-
-.backdrop {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.2);
-
-  overflow-y: scroll;
-
-  opacity: 1;
-
-  transition: opacity 250ms cubic-bezier(0.4, 0, 0.2, 1);
-}
-
-.backdrop.is-hidden {
-  opacity: 0;
-  pointer-events: none;
-}
-
-.modal {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%) scale(1);
-  background-color: #fff;
-
-  height: 568px;
-
-  display: flex;
-  padding: 30px 50px 30px 40px;
-
-  @include for-size(mobile) {
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%) scale(1);
-    width: 280px;
-
-    transition: transform 250ms $transition-function;
-
-    .backdrop.is-hidden & {
-      transform: translate(-50%, -50%) scale(1.2);
-    }
-  }
-
-  @include for-size(tablet) {
-    display: flex;
-    width: 618px;
-    padding: 42px 30px 40px 28px;
-  }
-
-  @include for-size(desktop) {
-    width: 882px;
-    padding: 50px 30px 40px;
-  }
-}
-
-.modal__img-wrap {
-  width: 240px;
-  height: 357px;
-  margin-bottom: 30px;
-
-  border-radius: 5px;
-
-  @include for-size(tablet) {
-    width: 264px;
-    height: 373px;
-    margin-right: 30px;
-  }
-
-  @include for-size(desktop) {
-    width: 396px;
-    height: 478px;
-  }
-}
-
-.modal__description {
-  max-width: 396px;
-}
-
-.modal__img {
-  margin-right: 30px;
-}
-
-.modal__title {
-  font-family: Roboto;
-  font-weight: 500;
-  font-size: 30px;
-  line-height: 35px;
-
-  color: #000000;
-  text-transform: uppercase;
-}
-
-.modal__ratings {
-  display: flex;
-  margin-top: 20px;
-  flex-direction: column;
-}
-
-.modal__descr-wrapper {
-  display: flex;
-  align-items: center;
-  margin-bottom: 10px;
-
-  &:last-child {
-    margin-bottom: 0;
-  }
-}
-
-.modal__criteria {
-  font-family: Roboto;
-  font-weight: 500;
-  font-size: 12px;
-  line-height: 16px;
-  color: #8c8c8c;
-}
-
-.modal__descr-wrapper > h3 {
-  font-family: Roboto;
-
-  font-weight: 700;
-  font-size: 12px;
-  line-height: 16px;
-
-  color: #000000;
-}
-
-.modal__descr-wrapper .orange {
-  width: 36px;
-  height: 16px;
-  border-radius: 5px;
-  background: #ff6b01;
-  text-align: center;
-  color: white;
-}
-
-.modal__descr-wrapper .grey {
-  width: 36px;
-  height: 16px;
-  border-radius: 5px;
-  background: #f7f7f7;
-  text-align: center;
-}
-
-.modal__criteria.a {
-  margin-right: 91px;
-}
-.modal__criteria.b {
-  margin-right: 97px;
-}
-.modal__criteria.c {
-  margin-right: 83px;
-}
-.modal__criteria.d {
-  margin-right: 120px;
-}
-
-.modal__genre-list {
-  display: flex;
-}
-
-.modal__genre-list li {
-  font-family: Roboto;
-  font-weight: 700;
-  font-size: 12px;
-  line-height: 16px;
-
-  color: #000000;
-}
-
-.modal__descr-title {
-  margin-top: 20px;
-  font-family: Roboto;
-
-  font-weight: 700;
-  font-size: 12px;
-  line-height: 16px;
-
-  text-transform: uppercase;
-
-  color: #000000;
-}
-
-.modal__descr-information {
-  font-family: Roboto;
-  font-weight: 700;
-  font-size: 12px;
-  line-height: 20px;
-
-  color: #000000;
-}
-
-.modal__buttons {
-  margin-top: 20px;
-}
-.modal__btn {
-  width: 136px;
-  height: 44px;
-  border-radius: 5px;
-
-  font-family: Roboto;
-  font-weight: 500;
-  font-size: 12px;
-  line-height: 16px;
-  text-align: center;
-  text-transform: uppercase;
-  cursor: pointer;
-}
-
-.modal__btn.watched {
-  background: #ff6b01;
-  border: none;
-  color: #ffffff;
-  opacity: 0.9;
-  &:hover {
-    opacity: 1;
-  }
-}
-
-.modal__btn.queue {
-  border: 1px solid #000000;
-  background: #ffffff;
-  color: #000000;
-  font-weight: 700;
-
-  &:hover {
-    background: #f7f7f7;
-  }
-}
