--- src/scss/components/_movie.scss
@@ -1,133 +1,0 @@
-.movie {
-  padding-top: 20px;
-  padding-bottom: 40px;
-
-  @include for-size(tablet) {
-    padding-top: 60px;
-    padding-bottom: 60px;
-  }
-}
-
-.movie__list {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 40px;
-
-  @include for-size(tablet) {
-    flex-direction: row;
-    flex-wrap: wrap;
-    margin-bottom: 50px;
-  }
-}
-
-.poster {
-  width: 280px;
-  margin-bottom: 20px;
-
-  cursor: pointer;
-
-  &:last-child {
-    margin-bottom: 0;
-  }
-
-  @include for-size(tablet) {
-    width: calc((100% - 30px) / 2);
-    margin-right: 30px;
-    margin-bottom: 30px;
-
-    &:nth-child(2n) {
-      margin-right: 0;
-    }
-
-    &:nth-last-child(-n + 2) {
-      margin-bottom: 0;
-    }
-  }
-
-  @include for-size(desktop) {
-    width: calc((100% - 2 * 30px) / 3);
-    /* margin-right: 30px; */
-
-    &:nth-child(2n) {
-      margin-right: 30px;
-    }
-
-    &:nth-child(3n) {
-      margin-right: 0;
-    }
-
-    &:nth-last-child(-n + 3) {
-      margin-bottom: 0;
-    }
-  }
-}
-
-.poster__img-wrap {
-  width: 100%;
-  height: 398px;
-  margin-bottom: 10px;
-
-  border-radius: 5px;
-}
-
-.poster__movie-info-wrap {
-  display: flex;
-  justify-content: baseline;
-  color: $orange-color;
-}
-
-.poster__movie-name {
-  @include font(12px, 1.33);
-  text-transform: uppercase;
-}
-
-.poster__genre-list {
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.poster__genre-item {
-  & > span:last-child {
-    display: none;
-  }
-
-  &:last-child > span:first-child {
-    display: none;
-  }
-
-  &:last-child > span:last-child {
-    display: inline-block;
-  }
-}
-
-.poster__movie-year {
-  margin: 0;
-  margin-left: 6px;
-}
-
-/* ========================== */
-
-.error {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-.error img {
-  margin-bottom: 20px;
-}
-.gallery__failure {
-  display: block;
-  margin: 0 auto;
-
-  font-family: Roboto;
-  font-weight: 500;
-  font-size: 26px;
-  text-align: center;
-
-  color: #000;
-
-  & > span {
-    color: #ff6b08;
-  }
-}
