+++ src/scss/global/_common.scss
@@ -10,13 +10,8 @@ html {
 body {
   background-color: $white-color;
   font-family: Roboto, sans-serif;
-  overflow: auto; //зачем? стоит по умолчанию
-  padding-top: 0px !important; // импортант нельзя ставить ЧИТАЙТЕ КОНСПЕКТ. Зачем паддинг топ, если в боди он отсутствует??
-
-  //добавил
-  font-size: 12px;
-  font-weight: 500;
-  line-height: 1.33;
+  overflow: auto;
+  padding-top: 0px !important;
 }
 .container {
   width: 100%;
@@ -47,60 +42,12 @@ body {
 img {
   display: block;
   width: 100%;
-  height: 100%;
+  height: auto;
 }
 a {
   font-style: none;
   text-decoration: none;
-
-  //добавил
-  display: inline-block;
 }
 svg {
   vertical-align: baseline;
-}
-
-//добавил
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-  margin: 0;
-
-  color: $black-color;
-}
-
-.list {
-  margin: 0;
-  padding: 0;
-  list-style: none;
-}
-
-.button {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-wrap: wrap;
-
-  border: 1px solid $black-color;
-  border-radius: 5px;
-  background-color: transparent;
-
-  text-transform: uppercase;
-}
-
-.visually-hidden {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  margin: -1px;
-  border: 0;
-  padding: 0;
-
-  white-space: nowrap;
-  clip-path: inset(100%);
-  clip: rect(0 0 0 0);
-  overflow: hidden;
 }
