--- src/scss/utils/_mixins.scss
@@ -1,33 +1,0 @@
-@mixin for-size($range) {
-  $mobile: 320px;
-  $tablet: 768px;
-  $desktop: 1024px;
-
-  @if $range==mobile {
-    @media (min-width: $mobile) {
-      @content;
-    }
-  } @else if $range==mobile-only {
-    @media (max-width: #{$tablet - 1}) {
-      @content;
-    }
-  } @else if $range==tablet {
-    @media (min-width: $tablet) {
-      @content;
-    }
-  } @else if $range==tablet-only {
-    @media (max-width: #{$desktop - 1}) {
-      @content;
-    }
-  } @else if $range==desktop {
-    @media (min-width: $desktop) {
-      @content;
-    }
-  }
-}
-
-@mixin font($fs, $lh, $fw: 500) {
-  font-size: $fs;
-  line-height: $lh;
-  font-weight: $fw;
-}
