+++ src/scss/utils/_variables.scss
@@ -7,5 +7,3 @@ $footer-text-color: #545454;
 $footer-background-color: #f7f7f7;
 $modal-name-color: #8c8c8c;
 $red-color: #ff001b;
-//добавил
-$transition-function: cubic-bezier(0.4, 0, 0.2, 1);
