+++ src/templates/header/spinner.hbs
@@ -1,3 +1,3 @@
-<div class="spinner-border" role="status" id="spinner">
+<div class="spinner-border" role="status">
   <span class="sr-only">Loading...</span>
 </div>
